refactor(Ladder): 사다리 생성 로직 LadderMaker -> Ladder 객체로 변경
